\name{makeWeightsAIC}
\alias{makeWeightsAIC}
\title{
A function to calculate AIC weights
}
\description{
This function calculates AIC weights for use in model selection and model averaging.
}
\usage{
makeWeightsAIC(aics)
}
\arguments{
  \item{aics}{
a (non-empty) numeric vector of AIC values.
}
}
\details{
FIXME - citation
}
\value{
Returns a vector of AIC weights.
}
\references{
FIXME - refs
}
\examples{
#not run, internal function
}