\name{school_district_bins}
\alias{school_district_bins}
\docType{data}
\title{
A data set containing the school district data.
}
\description{
This is a list where each entry is a different year.
}
\usage{data(state_bins)}
\format{
#FIXME - add in format
}
\examples{
data(school_district_bins)
head(school_district_bins[[1]])
names(school_district_bins)
}
